# docs/components/backend/analytics/openapi.json
{
  "components": {
    "schemas": {
...
      "ExplainRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MetricSnapshot"
          }
        ]
      },
      "ExplainResponse": {
        "properties": {
          "model": {
            "description": "The model that produced it.",
            "type": "string"
          },
          "person_context_entries": {
            "description": "How many of the caller's own entries fed the prompt.",
            "minimum": 0,
            "type": "integer"
          },
          "tenant_context_entries": {
            "description": "How many organisation entries fed the prompt.",
            "minimum": 0,
            "type": "integer"
          },
          "text": {
            "description": "The answer, as plain prose.",
            "type": "string"
          }
        },
        "required": [
          "text",
          "model",
          "tenant_context_entries",
          "person_context_entries"
        ],
        "type": "object"
      },
...
      "Grain": {
        "description": "Bucket width. The set is closed and validated server-side: the view is a\n`merge()` over every bronze database, so an operator-supplied interval\nexpression would be a direct route into that scan.",
        "enum": [
          "15m",
          "1s"
        ],
        "type": "string"
      },
...
      "IngestionIntensityResponse": {
        "properties": {
          "from": {
            "type": "string"
          },
          "grain": {
            "$ref": "#/components/schemas/Grain",
            "description": "Echoed resolved, not as asked: the caller may have pinned neither bound."
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/IngestionPoint"
            },
            "type": "array"
          },
          "scope": {
            "description": "The `source_database` the read was scoped to; absent when org-wide.",
            "type": [
              "string",
              "null"
            ]
          },
          "series": {
            "$ref": "#/components/schemas/Series"
          },
          "to": {
            "type": "string"
          },
          "truncated": {
            "description": "The group cap clipped the tail: the window is too wide for this grain\nand series to plot honestly. Never silently true — the UI says so.",
            "type": "boolean"
          }
        },
        "required": [
          "grain",
          "series",
          "from",
          "to",
          "truncated",
          "points"
        ],
        "type": "object"
      },
      "IngestionPoint": {
        "properties": {
          "bucket": {
            "description": "Bucket start as `YYYY-MM-DD HH:MM:SS`, always UTC — the reader's own\nzone would re-cut buckets the server already decided.",
            "type": "string"
          },
          "key": {
            "description": "Connector slug, stream name, or `all`, per the resolved `series`.",
            "type": "string"
          },
          "rows": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "bucket",
          "key",
          "rows"
        ],
        "type": "object"
      },
...
      "MetricSnapshot": {
        "description": "The reading as the viewer sees it, handed to the model as the thing to explain.",
        "properties": {
          "bucket_starts": {
            "description": "Bucket start dates every series is indexed by, oldest first.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "delta": {
            "description": "The tile's own change line, empty when it has none.",
            "type": "string"
          },
          "help": {
            "description": "The catalog's description of the metric, empty when it has none.",
            "type": "string"
          },
          "label": {
            "description": "The label the tile shows.",
            "type": "string"
          },
          "metric_key": {
            "description": "Catalog key, e.g. `tasks.closed`.",
            "type": "string"
          },
          "peer": {
            "description": "The tile's peer-comparison line, empty when it has none.",
            "type": "string"
          },
          "period": {
            "description": "What the period is called on screen, e.g. `month`.",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/SnapshotScope",
            "description": "Whose reading this is. Absent means one person's."
          },
          "series": {
            "description": "The chart's lines, when the reading is a chart rather than a tile.",
            "items": {
              "$ref": "#/components/schemas/SnapshotSeries"
            },
            "type": "array"
          },
          "since": {
            "description": "Inclusive start of the window, `YYYY-MM-DD`.",
            "type": "string"
          },
          "trend": {
            "description": "The sparkline's readings, oldest first.",
            "items": {
              "format": "double",
              "type": [
                "number",
                "null"
              ]
            },
            "type": "array"
          },
          "until": {
            "description": "Inclusive end of the window, `YYYY-MM-DD`.",
            "type": "string"
          },
          "value": {
            "description": "The formatted value the tile shows.",
            "type": "string"
          }
        },
        "required": [
          "metric_key",
          "label",
          "value",
          "period",
          "since",
          "until"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem+json. `context` varies by error category.",
        "properties": {
          "context": {
            "type": "object"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "context"
        ],
        "type": "object"
      },
...
      "Series": {
        "description": "What one plotted band counts. `Total` exists because the full-period trend\nplots a single line: grouping it by connector would multiply 400 days of\n15-minute buckets by the connector count for a series the chart then sums\nback down anyway.",
        "enum": [
          "connector",
          "stream",
          "total"
        ],
        "type": "string"
      },
      "SnapshotScope": {
        "description": "Whose reading this is.",
        "enum": [
          "person",
          "organisation"
        ],
        "type": "string"
      },
      "SnapshotSeries": {
        "description": "One line of a chart, as it is drawn.",
        "properties": {
          "label": {
            "type": "string"
          },
          "points": {
            "description": "Readings per bucket, oldest first; a gap is null.",
            "items": {
              "format": "double",
              "type": [
                "number",
                "null"
              ]
            },
            "type": "array"
          }
        },
        "required": [
          "label",
          "points"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-only query service over ClickHouse. Serves declarative metric definitions, computed metric results and their evidence, plus tenant-authored saved queries. The API Gateway mounts this service at /api/analytics.",
    "title": "Analytics API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ai/explain": {
      "post": {
        "operationId": "analytics_api.ai.explain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "description": "The tile to explain",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            },
            "description": "The explanation"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Explain one metric reading"
      }
    },
...
    "/v1/ingestion/intensity": {
      "get": {
        "operationId": "analytics_api.ingestion.intensity",
        "parameters": [
          {
            "description": "Bucket width: 15m or 1s",
            "in": "query",
            "name": "grain",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Bronze database to scope to, e.g. bronze_bamboohr",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "What one band counts: connector, stream or total",
            "in": "query",
            "name": "series",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Inclusive lower bound, RFC 3339",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Exclusive upper bound, RFC 3339",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionIntensityResponse"
                }
              }
            },
            "description": "Extraction intensity buckets"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Bronze extraction intensity per bucket"
      }
    },
...
  },
...
}